+++ src/components/Navigation/Tabs/LinkTab/index.ts
@@ -1,0 +1,4 @@
+import LinkTab from "./LinkTab";
+
+export default LinkTab;
+export * from "./LinkTab";