+++ src/main.ts
@@ -7,6 +7,8 @@ export { default as IfFeatureEnable } from "@/components/Layout/IfFeatureEnable"
 export * from "@/components/Layout/IfFeatureEnable";
 export { default as Logo } from "@/components/DataDisplay/Logo";
 export * from "@/components/DataDisplay/Logo";
+export { default as LinkTab } from "@/components/Navigation/Tabs/LinkTab";
+export * from "@/components/Navigation/Tabs/LinkTab";
 
 // Context
 export { default as ThemeProvider } from "@/context/Theme/ThemeProvider";