+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # [Versions](https://github.com/Tracktor/design-system-tracktor/releases)
 
-## v1.2.2
-- Update theme shape border radius
+## v1.3.0
+- Upgrade storybook
+- Add `LinkTab` component
+- Fix handle change tab event on `Tabs` stories